+++ packages/core/src/diff/schema.ts
@@ -84,17 +84,32 @@ function changesInSchema(
   newSchema: GraphQLSchema,
   addChange: AddChange,
 ) {
+  const defaultNames = {
+    query: 'Query',
+    mutation: 'Mutation',
+    subscription: 'Subscription',
+  };
   const oldRoot = {
-    query: (oldSchema.getQueryType() || ({} as GraphQLObjectType)).name,
-    mutation: (oldSchema.getMutationType() || ({} as GraphQLObjectType)).name,
-    subscription: (oldSchema.getSubscriptionType() || ({} as GraphQLObjectType))
-      .name,
+    query:
+      (oldSchema.getQueryType() || ({} as GraphQLObjectType)).name ??
+      defaultNames.query,
+    mutation:
+      (oldSchema.getMutationType() || ({} as GraphQLObjectType)).name ??
+      defaultNames.mutation,
+    subscription:
+      (oldSchema.getSubscriptionType() || ({} as GraphQLObjectType)).name ??
+      defaultNames.subscription,
   };
   const newRoot = {
-    query: (newSchema.getQueryType() || ({} as GraphQLObjectType)).name,
-    mutation: (newSchema.getMutationType() || ({} as GraphQLObjectType)).name,
-    subscription: (newSchema.getSubscriptionType() || ({} as GraphQLObjectType))
-      .name,
+    query:
+      (newSchema.getQueryType() || ({} as GraphQLObjectType)).name ??
+      defaultNames.query,
+    mutation:
+      (newSchema.getMutationType() || ({} as GraphQLObjectType)).name ??
+      defaultNames.mutation,
+    subscription:
+      (newSchema.getSubscriptionType() || ({} as GraphQLObjectType)).name ??
+      defaultNames.subscription,
   };
 
   if (isNotEqual(oldRoot.query, newRoot.query)) {